requirements.txt
acres==0.3.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
billiard==4.2.1
cachetools==5.5.2
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
ci-info==0.3.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.9
configparser==7.2.0
distro==1.9.0
dj-database-url==2.3.0
Django==5.2.1
django-autocomplete-light==3.12.1
django-filter==25.1
django-storages==1.14.6
django-widget-tweaks==1.5.0
etelemetry==0.3.1
filelock==3.18.0
google-api-core==2.25.0rc1
google-auth==2.40.2
google-cloud-core==2.4.3
google-cloud-storage==3.1.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
idna==3.10
isodate==0.6.1
jiter==0.10.0
joblib==1.5.0
kombu==5.5.3
looseversion==1.3.0
lxml==5.4.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.10.0
numpy==2.2.6
openai==1.82.0
packaging==25.0
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pgvector==0.4.1
pillow==11.2.1
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==6.31.0
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.29
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.5
pydantic_core==2.33.2
pydot==4.0.0
PyMuPDF==1.26.0
pyparsing==3.2.3
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyxnat==1.6.3
rdflib==6.3.2
redis==6.1.0
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.3
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
